Add license headers and refactor
