+++ components/wso2is.key.manager.tokenpersistence/src/main/java/org/wso2/is/key/manager/tokenpersistence/internal/TokenPersistenceServiceComponent.java
@@ -37,8 +37,8 @@ import org.wso2.carbon.registry.core.service.RegistryService;
 import org.wso2.carbon.registry.core.service.TenantRegistryLoader;
 import org.wso2.carbon.user.core.service.RealmService;
 import org.wso2.carbon.utils.ConfigurationContextService;
-import org.wso2.is.key.manager.tokenpersistence.adapter.ExtendedIndirectTokenRevocationHandlerImpl;
 import org.wso2.is.key.manager.tokenpersistence.dao.ExtendedAccessTokenDAOImpl;
+import org.wso2.is.key.manager.tokenpersistence.handler.ExtendedIndirectTokenRevocationHandlerImpl;
 import org.wso2.is.key.manager.tokenpersistence.listner.APIMOAuthApplicationMgtListener;
 import org.wso2.is.key.manager.tokenpersistence.processor.InMemoryOAuth2RevocationProcessor;
 import org.wso2.is.key.manager.tokenpersistence.processor.InMemoryRefreshTokenGrantProcessor;
